pyproject.toml
[project]
name = "whisperkey"
version = "0.2.0"
description = "Whisper key"
readme = "README.md"
license = "MIT"
authors = [
    { name = "Daniel", email = "danielratmiroff@pm.me" }
]
requires-python = ">=3.12"
dependencies = [
    "appdirs>=1.4.4",
    "dbus-python>=1.4.0",
    "notify2>=0.3.1",
    "openai>=2.15.0",
    "pillow>=10.0.0",
    "psutil>=7.2.1",
    "pyaudio>=0.2.14",
    "pydantic>=2.12.5",
    "pynput>=1.8.1",
    "pyperclip>=1.11.0",
    "pystray>=0.19.5",
]

[project.scripts]
whisperkey = "whisperkey.main:main"

[project.urls]
Homepage = "https://github.com/Danielratmiroff/whisper-key"
Repository = "https://github.com/Danielratmiroff/whisper-key"
Issues = "https://github.com/Danielratmiroff/whisper-key/issues"
Documentation = "https://github.com/Danielratmiroff/whisper-key#readme"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
